requirement.txt
opencv-python
pytesseract
numpy
pillow
selenium
scikit-learn
psutil
webdriver-manager
